+++ package.json
@@ -14,7 +14,7 @@
     "@types/uuid": "^3.4.3",
     "axios": "^0.21.1",
     "date-fns": "^1.29.0",
-    "decentraland-connect": "^2.6.0",
+    "decentraland-connect": "^2.7.0",
     "decentraland-ui": "^3.1.0",
     "flat": "^4.1.0",
     "react-intl": "^3.12.0",
