--- src/containers/EtherscanLink/index.ts
@@ -1,2 +1,0 @@
-import EtherscanLink from './EtherscanLink.container'
-export default EtherscanLink
