+++ src/containers/TransactionLink/TransactionLink.container.ts
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux'
 import { getChainId } from '../../modules/wallet/selectors'
-import { MapStateProps, MapDispatchProps } from './EtherscanLink.types'
-import EtherscanLink from './EtherscanLink'
+import { MapStateProps, MapDispatchProps } from './TransactionLink.types'
+import TransactionLink from './TransactionLink'
 
 const mapState = (state: any): MapStateProps => ({
   chainId: getChainId(state)
@@ -12,4 +12,4 @@ const mapDispatch = (_: any): MapDispatchProps => ({})
 export default connect(
   mapState,
   mapDispatch
-)(EtherscanLink)
+)(TransactionLink)
