+++ src/containers/TransactionLink/index.ts
@@ -1,0 +1,2 @@
+import TransactionLink from './TransactionLink.container'
+export default TransactionLink
