+++ src/containers/index.ts
@@ -1,4 +1,4 @@
-export { default as EtherscanLink } from './EtherscanLink'
+export { default as TransactionLink } from './TransactionLink'
 export { default as Footer } from './Footer'
 export { default as Navbar } from './Navbar'
 export { default as Profile } from './Profile'
